feat: show notification for random number seed changes
